backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# STT - Faster Whisper
faster-whisper==1.0.1

# Vision - YOLOv8
ultralytics==8.1.0
opencv-python-headless==4.9.0.80
pillow==10.2.0
numpy==1.26.3

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
